add express backend and refactor agent
