+++ package.json
@@ -3,7 +3,8 @@
   "version": "1.0.0",
   "main": "index.js",
   "scripts": {
-    "start": "node src/main.js"
+    "start": "node src/main.js",
+    "server": "node src/server.js"
   },
   "author": "",
   "license": "ISC",
@@ -13,7 +14,10 @@
     "chalk": "^5.4.1",
     "dotenv": "^16.5.0",
     "llamaindex": "^0.10.6",
-    "zod": "^3.24.4"
+    "zod": "^3.24.4",
+    "express": "^4.18.3",
+    "cors": "^2.8.5",
+    "morgan": "^1.10.0"
   },
   "type": "module"
 }
