+++ src/agent.js
@@ -1,0 +1,89 @@
+import { tool, agent } from "llamaindex";
+import { Ollama } from "@llamaindex/ollama";
+import { z } from "zod";
+import { Estudiantes } from "./lib/estudiantes.js";
+
+const DEBUG = true;
+
+const estudiantes = new Estudiantes();
+estudiantes.cargarEstudiantesDesdeJson();
+
+const systemPrompt = `
+Sos un asistente para gestionar estudiantes.
+Tu tarea es ayudar a consultar o modificar una base de datos de alumnos.
+
+Usá las herramientas disponibles para:
+- Buscar estudiantes por nombre o apellido
+- Agregar nuevos estudiantes
+- Mostrar la lista completa de estudiantes
+
+Respondé de forma clara y breve.
+`.trim();
+
+const ollamaLLM = new Ollama({
+    model: "qwen3:1.7b",
+    temperature: 0.75,
+    timeout: 2 * 60 * 1000,
+});
+
+const buscarPorNombreTool = tool({
+    name: "buscarPorNombre",
+    description: "Usa esta función para encontrar estudiantes por su nombre",
+    parameters: z.object({
+        nombre: z.string().describe("El nombre del estudiante a buscar"),
+    }),
+    execute: ({ nombre }) => {
+        const encontrados = estudiantes.buscarEstudiantePorNombre(nombre);
+        if (encontrados.length === 0) {
+            return `No se encontraron estudiantes con el nombre ${nombre}`;
+        }
+        return JSON.stringify(encontrados, null, 2);
+    },
+});
+
+const buscarPorApellidoTool = tool({
+    name: "buscarPorApellido",
+    description: "Usa esta función para encontrar estudiantes por su apellido",
+    parameters: z.object({
+        apellido: z.string().describe("El apellido del estudiante a buscar"),
+    }),
+    execute: ({ apellido }) => {
+       const encontrados = estudiantes.buscarEstudiantePorApellido(apellido);
+       if (encontrados.length === 0) {
+           return `No se encontraron estudiantes con el apellido ${apellido}`;
+       }
+       return JSON.stringify(encontrados, null, 2);
+    },
+});
+
+const agregarEstudianteTool = tool({
+    name: "agregarEstudiante",
+    description: "Usa esta función para agregar un nuevo estudiante",
+    parameters: z.object({
+        nombre: z.string().describe("El nombre del estudiante"),
+        apellido: z.string().describe("El apellido del estudiante"),
+        curso: z.string().describe("El curso del estudiante (ej: 4A, 4B, 5A)"),
+    }),
+    execute: ({ nombre, apellido, curso }) => {
+        estudiantes.agregarEstudiante(nombre, apellido, curso);
+        return `Estudiante ${nombre} ${apellido} agregado al curso ${curso}`;
+    },
+});
+
+const listarEstudiantesTool = tool({
+    name: "listarEstudiantes",
+    description: "Usa esta función para mostrar todos los estudiantes",
+    parameters: z.object({}),
+    execute: () => {
+        return JSON.stringify(estudiantes.listarEstudiantes(), null, 2);
+    },
+});
+
+const elAgente = agent({
+    tools: [buscarPorNombreTool, buscarPorApellidoTool, agregarEstudianteTool, listarEstudiantesTool],
+    llm: ollamaLLM,
+    verbose: DEBUG,
+    systemPrompt: systemPrompt,
+});
+
+export { elAgente };
