+++ src/main.js
@@ -1,93 +1,6 @@
-import { tool, agent } from "llamaindex";
-import { Ollama } from "@llamaindex/ollama";
-import { z } from "zod";
 import { empezarChat } from "./lib/cli-chat.js";
-import { Estudiantes } from "./lib/estudiantes.js";
+import { elAgente } from "./agent.js";
 
-// Configuración
-const DEBUG = true;
-
-// Instancia de la clase Estudiantes
-const estudiantes = new Estudiantes();
-estudiantes.cargarEstudiantesDesdeJson();
-
-// System prompt básico
-const systemPrompt = `
-Sos un asistente para gestionar estudiantes.
-Tu tarea es ayudar a consultar o modificar una base de datos de alumnos.
-
-Usá las herramientas disponibles para:
-- Buscar estudiantes por nombre o apellido
-- Agregar nuevos estudiantes
-- Mostrar la lista completa de estudiantes
-
-Respondé de forma clara y breve.
-`.trim();
-
-const ollamaLLM = new Ollama({
-    model: "qwen3:1.7b",
-    temperature: 0.75,
-    timeout: 2 * 60 * 1000, // Timeout de 2 minutos
-});
-
-
-// TODO: Implementar la Tool para buscar por nombre
-const buscarPorNombreTool = tool({
-    name: "buscarPorNombre",
-    description: "Usa esta función para encontrar estudiantes por su nombre",
-    parameters: z.object({
-        nombre: z.string().describe("El nombre del estudiante a buscar"),
-    }),
-    execute: ({ nombre }) => {
-        // Tu código aquí
-    },
-});
-
-// TODO: Implementar la Tool para buscar por apellido
-const buscarPorApellidoTool = tool({
-    name: "buscarPorApellido",
-    description: "Usa esta función para encontrar estudiantes por su apellido",
-    parameters: z.object({
-        apellido: z.string().describe("El apellido del estudiante a buscar"),
-    }),
-    execute: ({ apellido }) => {
-       return;
-    },
-});
-
-// TODO: Implementar la Tool para agregar estudiante
-const agregarEstudianteTool = tool({
-    name: "agregarEstudiante",
-    description: "Usa esta función para agregar un nuevo estudiante",
-    parameters: z.object({
-        nombre: z.string().describe("El nombre del estudiante"),
-        apellido: z.string().describe("El apellido del estudiante"),
-        curso: z.string().describe("El curso del estudiante (ej: 4A, 4B, 5A)"),
-    }),
-    execute: ({ nombre, apellido, curso }) => {
-        return;
-    },
-});
-
-// TODO: Implementar la Tool para listar estudiantes
-const listarEstudiantesTool = tool({
-    name: "listarEstudiantes",
-    description: "Usa esta función para mostrar todos los estudiantes",
-    parameters: z.object({}),
-    execute: () => {
-        return;
-    },
-});
-
-// Configuración del agente
-const elAgente = agent({
-    tools: [buscarPorNombreTool, buscarPorApellidoTool, agregarEstudianteTool, listarEstudiantesTool],
-    llm: ollamaLLM,
-    verbose: DEBUG,
-    systemPrompt: systemPrompt,
-});
-
-// Mensaje de bienvenida
 const mensajeBienvenida = `
 ¡Hola! Soy tu asistente para gestionar estudiantes.
 Puedo ayudarte a:
@@ -98,5 +11,4 @@ Puedo ayudarte a:
 ¿Qué necesitás?
 `;
 
-// Iniciar el chat
 empezarChat(elAgente, mensajeBienvenida);
